move types to node package
