+++ plugins/scaffolder-node/src/checkpoints/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2024 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,4 @@
  * limitations under the License.
  */
 
-/**
- * The scaffolder-node module for `@backstage/plugin-scaffolder-backend`.
- *
- * @packageDocumentation
- */
-
-export * from './actions';
-export * from './tasks';
-export * from './files';
 export * from './types';
-export * from './checkpoints';
