+++ plugins/scaffolder-node/src/tasks/types.ts
@@ -16,7 +16,8 @@
 
 import { BackstageCredentials } from '@backstage/backend-plugin-api';
 import { TaskSpec } from '@backstage/plugin-scaffolder-common';
-import { JsonObject, JsonValue, Observable } from '@backstage/types';
+import { JsonObject, Observable } from '@backstage/types';
+import { UpdateCheckpointOptions } from '../checkpoints';
 
 /**
  * TaskSecrets
@@ -129,19 +130,7 @@ export interface TaskContext {
     | undefined
   >;
 
-  updateCheckpoint?(
-    options:
-      | {
-          key: string;
-          status: 'success';
-          value: JsonValue;
-        }
-      | {
-          key: string;
-          status: 'failed';
-          reason: string;
-        },
-  ): Promise<void>;
+  updateCheckpoint?(options: UpdateCheckpointOptions): Promise<void>;
 
   serializeWorkspace?(options: { path: string }): Promise<void>;
